feature/OP-898 - Search Results CSV

+++ app/upload/views.py
@@ -58,7 +58,7 @@ def post(request_id):
     files = request.files
     file_ = files[next(files.keys())]
     filename = secure_filename(file_.filename)
-    is_update = eval_request_bool(request.form.get('update'), False)
+    is_update = eval_request_bool(request.form.get('update'))
     if not is_update and upload_exists(request_id, filename):
         response = {
             "files": [{
@@ -166,13 +166,13 @@ def delete(r_id_type, r_id, filecode):
                 r_id = response.request_id
 
             path = ''
-            quarantined_only = eval_request_bool(request.form.get('quarantined_only'), False)
+            quarantined_only = eval_request_bool(request.form.get('quarantined_only'))
             if quarantined_only:
                 path = os.path.join(
                     current_app.config['UPLOAD_QUARANTINE_DIRECTORY'],
                     r_id
                 )
-            elif eval_request_bool(request.form.get('updated_only'), False):
+            elif eval_request_bool(request.form.get('updated_only')):
                 path = os.path.join(
                     current_app.config['UPLOAD_DIRECTORY'],
                     r_id,
@@ -191,13 +191,17 @@ def delete(r_id_type, r_id, filecode):
                         r_id
                     )
             filepath = os.path.join(path, filename)
+            found = False
             if path != '':
                 if quarantined_only:
-                    os.path.exists(filepath)
-                    os.remove(filepath)
+                    if os.path.exists(filepath):
+                        os.remove(filepath)
+                        found = True
                 else:
-                    fu.exists(filepath)
-                    fu.remove(filepath)
+                    if fu.exists(filepath):
+                        fu.remove(filepath)
+                        found = True
+            if found:
                 response = {"deleted": filename}
             else:
                 response = {"error": "Upload not found."}
@@ -227,7 +231,7 @@ def status():
             get_upload_key(
                 request.args['request_id'],
                 secure_filename(request.args['filename']),
-                eval_request_bool(request.args.get('for_update'), False)
+                eval_request_bool(request.args.get('for_update'))
             )
         )
         if status is not None:
